+++ docker-compose.yml
@@ -14,11 +14,15 @@ services:
       - "5432:5432"  # Maps port 5432 on the host to port 5432 in the container
     volumes:
       - db_data:/var/lib/postgresql/data  # Persistent data storage
+    healthcheck:
+      test: [ "CMD-SHELL", "pg_isready" ]
+      interval: 10s
+      timeout: 5s
+      retries: 5
 
   pgadmin:
     container_name: pgadmin4
     image: dpage/pgadmin4
-    restart: always
     environment:
       PGADMIN_DEFAULT_EMAIL: admin@example.com
       PGADMIN_DEFAULT_PASSWORD: admin123
@@ -27,22 +31,22 @@ services:
     depends_on:
       - db
 
-  trading_bot_api:
-    container_name: trading_bot_api
-    build: ./src
-    restart: always
-    ports:
-      - "8086:8086"
-    depends_on:
-      - db
+  start_infra:
+    build: ./src/infra
+    container_name: start_infra
     environment:
-      DB_HOST: db
+      DB_CONN: "postgresql://db_user:pgpassword123@db/opa_db"
+    depends_on:
+      db:
+          condition: service_healthy
 
   spark-master:
     build:
-      context: ./load_data
+      context: ./src/load_data_etl
       dockerfile: Dockerfile
-    command: bin/spark-class org.apache.spark.deploy.master.Master 
+    command: bin/spark-class org.apache.spark.deploy.master.Master
+    environment:
+      DOWNLOAD_YEAR: 2024
     ports:
       - "9090:8080"
       - "7077:7077"
@@ -74,6 +78,17 @@ services:
       SPARK_MASTER_URL: spark://spark-master:7077
     volumes:
       - ./load_data/data:/data
+
+  # trading_bot_api:
+  #   container_name: trading_bot_api
+  #   build: ./src
+  #   restart: always
+  #   ports:
+  #     - "8086:8086"
+  #   depends_on:
+  #     - db
+  #   environment:
+  #     DB_CONN: "postgresql://db_user:pgpassword123@db/opa_db"
 
 volumes:
   db_data:
