+++ src/infra/requirements.txt
@@ -1,0 +1,3 @@
+psycopg2-binary==2.9.9
+SQLAlchemy==2.0.30
+typing_extensions==4.12.2
