+++ src/load_data_etl/download_data.py
@@ -59,9 +59,8 @@ def load_historical_data_from_year(symbols, year):
         if not year:
             year = 2024
 
-        period = convert_to_date_object((datetime.today() - timedelta(days=1)).strftime('%Y-%m-%d')) - convert_to_date_object(
-            datetime.date(f"{year}-01-01"))
-        dates = pd.date_range(end=datetime.today(), periods=period.days + 1).to_pydatetime().tolist()
+        period = convert_to_date_object((datetime.today() - timedelta(days=1)).strftime('%Y-%m-%d')) - convert_to_date_object(f"{year}-01-01")
+        dates = pd.date_range(end=datetime.today(), periods=period.days).to_pydatetime().tolist()
         dates = [date.strftime("%Y-%m-%d") for date in dates]
 
         download_daily_klines(trading_type="spot",
