+++ src/trading_bot_api/binance_response_formatter.py
