+++ src/trading_bot_api/kpis.py
@@ -1,5 +1,4 @@
-from cross_cutting import db_driver
-from cross_cutting import technical_indicators as ti
+from src.trading_bot_api import db_driver, technical_indicators as ti
 import pandas as pd
 import api_settings as settings
