+++ src/trading_bot_api/start_db.py
@@ -1,0 +1,3 @@
+import db_driver
+
+db_driver.create_db()
